requirement.txt
Flask==2.0.3  # Flask for web framework
pyserial==3.5  # Serial communication with the Raspberry Pi Pico
ultralytics==8.3.19        # YOLOv8 for object detection
torch==2.5.0               # PyTorch for YOLOv8
torchvision==0.20.0        # TorchVision for YOLOv8
opencv-python==4.10.0.84   # OpenCV for camera integration or image processing

pip install -r requirements.txt ##This is to quickly install dependancies


